Moved Generics responsibility from DbContainer to BaseRepository
