+++ DataLayer/Repositories/FileTypeRepository.cs
@@ -1,0 +1,10 @@
+﻿namespace DataLayer
+{
+    public class FileTypeRepository : BaseRepository<FileType>, IFileTypeRepository
+    {
+        public FileTypeRepository(IDbContainer dbContainer) : base(dbContainer)
+        { }
+
+
+    }
+}
