+++ DataLayer/Repositories/IFileTypeRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace DataLayer
+{
+    public interface IFileTypeRepository : IBaseRepository<FileType>
+    {
+    }
+}
